# docs/api/openapi-v1.1.yaml
openapi: 3.0.3
info:
  title: Relationship Platform API
  version: 1.1.0
  description: 'Product-facing REST API contract for the MVP: Dating/Love + Friendship
    + Marriage. Contract aligned to the v1.1 schema: one shared partner-preference and
    compatibility model across all relationship modes, plus one optional lightweight
    spiritual profile. Community and analytics features not represented in the MVP
    schema are explicitly marked as deferred or decision-required in the accompanying
    decisions document.'
servers:
- url: http://localhost:8080/api/v1
  description: Local development
- url: https://dev-api.example.com/api/v1
  description: Development/staging placeholder
...
paths:
...
  /master-data/{categoryCode}:
    get:
      summary: List active values for a master-data category
      tags:
      - Master Data
      parameters:
      - name: categoryCode
        in: path
        required: true
        schema:
          type: string
      - name: search
        in: query
        schema:
          type: string
      - name: page
        in: query
        schema:
          type: integer
          default: 0
      - name: size
        in: query
        schema:
          type: integer
          default: 100
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '401':
          description: Authentication required or invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
  /master-data/categories:
    get:
      summary: List master-data categories
      tags:
      - Master Data
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '401':
          description: Authentication required or invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
...
  /media/upload-url:
    post:
      summary: Create a presigned upload URL
      tags:
      - Media
      security:
      - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - filename
              - mimeType
              - fileSizeBytes
              properties:
                filename:
                  type: string
                mimeType:
                  type: string
                fileSizeBytes:
                  type: integer
                purpose:
                  type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiResponse'
                - type: object
                  properties:
                    data:
                      type: object
                      additionalProperties: true
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '401':
          description: Authentication required or invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
  /media/complete:
    post:
      summary: Complete a media upload
      tags:
      - Media
      security:
      - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - mediaFileId
              properties:
                mediaFileId:
                  type: string
                  format: uuid
                checksum:
                  type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiResponse'
                - type: object
                  properties:
                    data:
                      $ref: '#/components/schemas/MediaFile'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '401':
          description: Authentication required or invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
  /media/{mediaId}:
    delete:
      summary: Delete owned media
      tags:
      - Media
      security:
      - bearerAuth: []
      parameters:
      - name: mediaId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '401':
          description: Authentication required or invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
        data: {}
        message:
          type: string
        timestamp:
          type: string
          format: date-time
    ApiErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          $ref: '#/components/schemas/ApiError'
        timestamp:
          type: string
          format: date-time
    ApiError:
      type: object
      required:
      - code
      - message
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: true
...
    MediaFile:
      type: object
      properties:
        id:
          type: string
          format: uuid
        originalFilename:
          type: string
        mimeType:
          type: string
        fileSizeBytes:
          type: integer
          format: int64
        url:
          type: string
          format: uri
        status:
          type: string
...