# shared/src/main/java/com/pavitraristaa/common/api/ApiResponse.java
package com.pavitraristaa.common.api;

import java.time.Instant;

public record ApiResponse<T>(boolean success, T data, String message, Instant timestamp) {

    public static <T> ApiResponse<T> ok(T data, String message) {
        return new ApiResponse<>(true, data, message, Instant.now());
    }

    public static ApiResponse<Void> ok(String message) {
        return new ApiResponse<>(true, null, message, Instant.now());
    }
}